+++ extensions/notion/src/components/index.tsx
@@ -1,2 +1,2 @@
-export { CreatePageForm } from "./forms/CreatePageForm";
+export { CreateDataBasePageForm } from "./forms/CreatePageForm";
 export { PageListItem } from "./PageListItem";